+++ src/main/java/com/web/doitcommit/domain/hashtagCategory/HashtagCategory.java
@@ -1,4 +1,4 @@
-package com.web.doitcommit.domain.hashtag;
+package com.web.doitcommit.domain.hashtagCategory;
 
 import lombok.*;
