+++ src/main/java/com/web/doitcommit/domain/hashtagCategory/HashtagCategoryRepository.java
@@ -1,10 +1,6 @@
-package com.web.doitcommit.domain.hashtag;
+package com.web.doitcommit.domain.hashtagCategory;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.Calendar;
-import java.util.List;
 
 public interface HashtagCategoryRepository extends JpaRepository<HashtagCategory, Long>, HashtagCategoryRepositoryQuerydsl {
