+++ src/main/java/com/web/doitcommit/domain/hashtagCategory/HashtagCategoryRepositoryQuerydsl.java
@@ -1,6 +1,4 @@
-package com.web.doitcommit.domain.hashtag;
-
-import org.springframework.data.jpa.repository.Query;
+package com.web.doitcommit.domain.hashtagCategory;
 
 import java.util.List;
