+++ src/main/java/com/web/doitcommit/dto/board/BoardOfMemberResDto.java
@@ -2,7 +2,7 @@ package com.web.doitcommit.dto.board;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.web.doitcommit.domain.board.Board;
-import com.web.doitcommit.domain.hashtag.BoardHashtag;
+import com.web.doitcommit.domain.boardHashtag.BoardHashtag;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
