+++ src/main/java/com/web/doitcommit/dto/board/MainViewBoardListResDto.java
@@ -1,7 +1,7 @@
 package com.web.doitcommit.dto.board;
 
 import com.web.doitcommit.domain.board.Board;
-import com.web.doitcommit.domain.hashtag.BoardHashtag;
+import com.web.doitcommit.domain.boardHashtag.BoardHashtag;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import java.util.ArrayList;
