+++ src/test/java/com/web/doitcommit/service/comment/CommentServiceImplTest.java
@@ -11,10 +11,7 @@ import com.web.doitcommit.domain.comment.MemberTag;
 import com.web.doitcommit.domain.comment.MemberTagRepository;
 import com.web.doitcommit.domain.files.MemberImage;
 import com.web.doitcommit.domain.files.MemberImageRepository;
-import com.web.doitcommit.domain.hashtag.BoardHashtag;
 import com.web.doitcommit.domain.hashtag.BoardHashtagRepository;
-import com.web.doitcommit.domain.hashtag.HashtagCategory;
-import com.web.doitcommit.domain.hashtag.HashtagCategoryRepository;
 import com.web.doitcommit.domain.member.AuthProvider;
 import com.web.doitcommit.domain.member.Member;
 import com.web.doitcommit.domain.member.MemberRepository;
